requirements.txt
pymupdf
pdfplumber
langgraph
pydantic
streamlit
python-dotenv
xai-sdk
httpx
